Split dra to add func_800FD7C0
