+++ config/splat.dra.yaml
@@ -39,6 +39,8 @@ segments:
       - [0x3B7A0, rodata]
       - [0x3BF10, rodata]
       - [0x3C160, rodata]
+      - [0x3CBC4, rodata]
+      - [0x3CBDC, rodata]
       - [0x3CC80, rodata]
       - [0x3CD00, rodata]
       - [0x3CD10, rodata]
@@ -54,6 +56,8 @@ segments:
       - [0x41F70, rodata]
       - [0x41FB0, rodata]
       - [0x42000, rodata]
-      - [0x42398, c]
+      - [0x42398, c, 42398]
+      - [0x5D7C0, c, 5D7C0]
+      - [0x5D874, c, 5D874]
       - [0x962AC, data]
   - [0x119870]
