+++ src/dra/5D7C0.c
@@ -1,0 +1,6 @@
+#include "common.h"
+#include "dra.h"
+#include "objects.h"
+#include "sfx.h"
+
+INCLUDE_ASM("asm/dra/nonmatchings/5D7C0", func_800FD7C0);
